renamed from password enforcer to lantern extensions
