+++ app/views/lantern_extensions/_plugin_settings.html.erb
