+++ lib/lantern_extensions.rb
@@ -1,0 +1,6 @@
+require "lantern_extensions/version"
+require "lantern_extensions/engine"
+require "lantern_extensions/nullable_path"
+
+module LanternExtensions
+end
