+++ lib/lantern_extensions/nullable_path.rb
@@ -1,8 +1,8 @@
-module PasswordEnforcer
+module LanternExtensions
   class NullablePath
     def to_str
-      base_path = File.join(Rails.root.to_s, 'gems', 'plugins', 'password_enforcer', 'messages')
-      if Canvas::Plugin.find(:password_enforcer).enabled?
+      base_path = File.join(Rails.root.to_s, 'gems', 'plugins', 'lantern_extensions', 'messages')
+      if Canvas::Plugin.find(:lantern_extensions).enabled?
         base_path
       else
         # this is a path that is known to not exist
