+++ lib/lantern_extensions/version.rb
@@ -1,0 +1,3 @@
+module LanternExtensions
+  VERSION = "0.1.0"
+end
