--- lib/password_enforcer.rb
@@ -1,6 +1,0 @@
-require "password_enforcer/version"
-require "password_enforcer/engine"
-require "password_enforcer/nullable_path"
-
-module PasswordEnforcer
-end
