--- lib/password_enforcer/version.rb
@@ -1,3 +1,0 @@
-module PasswordEnforcer
-  VERSION = "0.1.0"
-end
